start creating chain of fallback services for quotes, using hystrix.

+++ spring-nanotrader-data/src/main/java/org/springframework/nanotrader/data/service/QuoteService.java
@@ -39,7 +39,7 @@ public interface QuoteService {
 
 	public Quote saveQuote(Quote quote);
 
-	public Map<String, Long> marketSummary();
+	public Map<String, Float> marketSummary();
 
 	public List<Quote> topGainers();
